Add end line, implement section component for home page
