+++ e-shop/src/app/assets/icons/index.ts
@@ -2,12 +2,16 @@ import Search from './search.svg';
 import Cart from './cart.svg';
 import User from './user.svg';
 import Shipping from './shipping.svg';
+import Happy from './happy.svg';
+import Guarantee from './guarantee.svg';
 
 const Icon = {
   Search,
   Cart,
   User,
   Shipping,
+  Happy,
+  Guarantee
 };
 
 export default Icon;
