+++ e-shop/src/app/assets/images/index.ts
@@ -1,0 +1,11 @@
+import Category1 from './fashion-category-1.png';
+import Category2 from './fashion-category-2.png';
+import Category3 from './fashion-category-3.png';
+
+const Images = {
+  Category1,
+  Category2,
+  Category3
+};
+
+export default Images;
