--- e-shop/src/app/pages/home/Home.tsx
@@ -1,280 +1,0 @@
-import ProductList from '../../shared/components/ProductlList';
-import productData from '../../shared/constant/product-data';
-import Policy from './components/Policy';
-import Icon from '../../assets/icons';
-
-const Home = () => {
-  return (
-    <main>
-      <section className="section-banner page-section">
-        <div className="container">
-          <div className="banner-inner">
-            <div className="banner-content">
-              <h2 className="banner-title">
-                Sale of the
-                <span className="text-primary">summer</span> collection
-              </h2>
-              <a href="" className="banner-link">
-                <span className="btn btn-icon-primary">
-                  <svg
-                    width="17"
-                    height="14"
-                    viewBox="0 0 17 14"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      d="M9.61405 -9.53674e-07L15.7729 5.84095L17 7.00368L15.7729 8.16745L9.61405 14L8.38693 12.8446L13.6745 7.82609L0 7.82609V6.18231L13.6745 6.18231L8.38693 1.16588L9.61405 -9.53674e-07Z"
-                      fill="white"
-                    />
-                  </svg>
-                </span>
-                <span className="banner-link-text">SHOP NOW</span>
-              </a>
-            </div>
-            <ul className="offers-list">
-              <Policy
-                icon = {Icon.Shipping}
-                title = "Free Shipping"
-                content = "On purchases over $199"
-              />
-              <Policy
-                icon = {Icon.Shipping}
-                title = "99% Satisfied Customers"
-                content = "Our clients' opinions speak for themselves"
-              />
-              <Policy
-                icon = {Icon.Shipping}
-                title = "Originality Guaranteed"
-                content = "30 days warranty for each product from our store"
-              />
-            </ul>
-          </div>
-        </div>
-      </section>
-      <section className="section-category page-section">
-        <div className="container">
-          <ul className="row categories-list">
-            <li className="category-item col-6 col-sm-12">
-              <div className="category">
-                <div className="category-img">
-                  <img src="~/images/fashion-category-1.png" alt="New arrivalsare now in!" />
-                </div>
-
-                <div className="category-content">
-                  <h4 className="category-title">New arrivalsare now in!</h4>
-                  <a className="btn btn-secondary" href="">SHOW COLLECTION</a>
-                </div>
-              </div>
-            </li>
-            <li className="category-item col-3 col-sm-6">
-              <div className="category">
-                <div className="category-img">
-                  <img src="img/fashion-category-2.png" alt="Basic t-shirts" />
-                </div>
-                <div className="category-content">
-                  <h4 className="category-title">
-                    Basic t-shirts <br />
-                    $29,99
-                  </h4>
-                  <a className="btn btn-secondary" href="">MORE DETAILS</a>
-                </div>
-              </div>
-            </li>
-            <li className="category-item col-3 col-sm-6">
-              <div className="category">
-                <div className="category-img">
-                  <img src="img/fashion-category-3.png" alt="Sale this summer" />
-                </div>
-                <div className="category-content">
-                  <span className="badge badge-primary">-50%</span>
-                  <h4 className="category-title">
-                    Sale this <br />
-                    summer
-                  </h4>
-                  <a className="btn btn-secondary" href="">VIEW ALL</a>
-                </div>
-              </div>
-            </li>
-          </ul>
-        </div>
-      </section>
-      <section className="section-for-you page-section">
-        <div className="container">
-          <div className="section-heading">
-            <h3 className="section-title">Selected just for you</h3>
-            <button className="btn btn-outline-secondary">SHOW MORE</button>
-          </div>
-          <ProductList data={productData}/>
-        </div>
-      </section>
-      <section className="section-benefits page-section">
-        <div className="container">
-          <div className="section-heading">
-            <h3 className="section-title text-center">Why should you choose us?</h3>
-          </div>
-          <ul className="benefits-list row">
-            <li className="benefit-item col-3 col-sm-12">
-              <div className="media media-2 media-vertical">
-                <span className="media-img benefit-icon">
-                  <svg
-                    width="38"
-                    height="22"
-                    viewBox="0 0 38 22"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      d="M36.7111 9.24104L35.0405 5.40489C34.7437 4.72155 34.2499 4.1397 33.6208 3.73186C32.9917 3.32401 32.2551 3.10821 31.5028 3.11136H28.7984V2.13313C28.7979 1.74426 28.6414 1.37146 28.3632 1.09648C28.085 0.821506 27.7079 0.666798 27.3145 0.666284H8.44657C8.31637 0.66538 8.18727 0.690066 8.0668 0.738906C7.94632 0.787746 7.83687 0.859767 7.74479 0.950784C7.65272 1.0418 7.57986 1.15 7.53045 1.26909C7.48105 1.38818 7.45608 1.51579 7.45699 1.64451C7.46154 1.90254 7.56725 2.14876 7.75186 2.33124C7.93647 2.51373 8.18554 2.61824 8.44657 2.62273H26.7436V9.84332C26.7419 10.0562 26.783 10.2673 26.8646 10.4643C26.9463 10.6613 27.0667 10.8402 27.219 10.9908C27.3713 11.1413 27.5523 11.2604 27.7516 11.341C27.9509 11.4217 28.1644 11.4624 28.3798 11.4607H34.9639V16.0109C34.964 16.0553 34.9553 16.0993 34.9381 16.1404C34.921 16.1815 34.8958 16.2188 34.864 16.2502C34.8322 16.2816 34.7945 16.3065 34.7529 16.3235C34.7114 16.3404 34.6669 16.3491 34.6219 16.3489H33.4014C33.0806 15.8015 32.6203 15.3468 32.0663 15.0301C31.5123 14.7134 30.8839 14.5458 30.244 14.544C29.6056 14.5538 28.9804 14.7247 28.4278 15.0406C27.8752 15.3565 27.4134 15.8069 27.0865 16.3489H16.9283C16.6075 15.8015 16.1472 15.3468 15.5932 15.0301C15.0392 14.7134 14.4108 14.5458 13.7709 14.544C13.1326 14.5538 12.5074 14.7247 11.9547 15.0406C11.4021 15.3565 10.9404 15.8069 10.6134 16.3489H7.11403C6.98151 16.3448 6.84951 16.367 6.72586 16.4143C6.60221 16.4616 6.48943 16.533 6.39422 16.6242C6.299 16.7154 6.2233 16.8245 6.1716 16.9452C6.1199 17.0659 6.09326 17.1956 6.09326 17.3267C6.09326 17.4577 6.1199 17.5875 6.1716 17.7081C6.2233 17.8288 6.299 17.938 6.39422 18.0292C6.48943 18.1204 6.60221 18.1918 6.72586 18.239C6.84951 18.2863 6.98151 18.3085 7.11403 18.3044H10.1191C10.16 19.2348 10.5626 20.1137 11.2431 20.7578C11.9235 21.4019 12.8293 21.7614 13.7714 21.7614C14.7135 21.7614 15.6192 21.4019 16.2996 20.7578C16.9801 20.1137 17.3827 19.2348 17.4236 18.3044H26.5922C26.6331 19.2348 27.0357 20.1137 27.7162 20.7578C28.3966 21.4019 29.3024 21.7614 30.2444 21.7614C31.1865 21.7614 32.0923 21.4019 32.7727 20.7578C33.4532 20.1137 33.8558 19.2348 33.8967 18.3044H34.6956C35.3102 18.3019 35.8988 18.0594 36.3334 17.6298C36.7679 17.2003 37.0132 16.6184 37.0158 16.0109V10.7458C37.0066 10.2302 36.9034 9.72043 36.7111 9.24104ZM13.7709 19.8033C13.4405 19.8033 13.1175 19.7065 12.8427 19.525C12.568 19.3435 12.3539 19.0856 12.2274 18.7839C12.101 18.4821 12.0679 18.1501 12.1324 17.8297C12.1968 17.5094 12.3559 17.2151 12.5896 16.9842C12.8232 16.7532 13.1209 16.5959 13.445 16.5322C13.769 16.4685 14.1049 16.5012 14.4102 16.6262C14.7154 16.7512 14.9764 16.9629 15.1599 17.2344C15.3435 17.506 15.4415 17.8253 15.4415 18.1519C15.4369 18.5885 15.2594 19.0059 14.9471 19.3146C14.6348 19.6233 14.2125 19.7988 13.7709 19.8033ZM30.244 19.8033C29.9135 19.8033 29.5906 19.7065 29.3158 19.525C29.0411 19.3435 28.827 19.0856 28.7005 18.7839C28.5741 18.4821 28.541 18.1501 28.6055 17.8297C28.6699 17.5094 28.829 17.2151 29.0627 16.9842C29.2963 16.7532 29.594 16.5959 29.918 16.5322C30.2421 16.4685 30.578 16.5012 30.8833 16.6262C31.1885 16.7512 31.4494 16.9629 31.633 17.2344C31.8166 17.506 31.9146 17.8253 31.9146 18.1519C31.9099 18.5885 31.7325 19.0059 31.4202 19.3146C31.1079 19.6233 30.6856 19.7988 30.244 19.8033ZM28.7601 5.09889H31.4615C31.8185 5.09447 32.1685 5.19609 32.4664 5.39059C32.7643 5.58509 32.9962 5.86351 33.1321 6.1898L34.5806 9.50624H28.7601V5.09889Z"
-                      fill="#333333"
-                    />
-                    <path
-                      d="M5.59239 6.72151H12.5922C12.7247 6.72564 12.8567 6.7034 12.9804 6.65612C13.104 6.60884 13.2168 6.53747 13.312 6.44627C13.4072 6.35506 13.4829 6.24587 13.5346 6.12518C13.5863 6.00449 13.613 5.87476 13.613 5.74369C13.613 5.61262 13.5863 5.48289 13.5346 5.3622C13.4829 5.2415 13.4072 5.13231 13.312 5.04111C13.2168 4.9499 13.104 4.87854 12.9804 4.83126C12.8567 4.78397 12.7247 4.76174 12.5922 4.76587H5.59239C5.46218 4.76496 5.33309 4.78965 5.21261 4.83849C5.09214 4.88734 4.98268 4.95937 4.89061 5.05039C4.79854 5.14142 4.72568 5.24962 4.67627 5.36873C4.62686 5.48783 4.60189 5.61545 4.60281 5.74417C4.6076 6.00206 4.71343 6.24805 4.89801 6.43035C5.0826 6.61264 5.33152 6.71702 5.59239 6.72151Z"
-                      fill="#333333"
-                    />
-                    <path
-                      d="M3.72838 10.4818H10.7282C10.8607 10.4859 10.9927 10.4637 11.1163 10.4164C11.24 10.3691 11.3528 10.2978 11.448 10.2066C11.5432 10.1154 11.6189 10.0062 11.6706 9.88551C11.7223 9.76483 11.7489 9.63512 11.7489 9.50407C11.7489 9.37301 11.7223 9.2433 11.6706 9.12262C11.6189 9.00194 11.5432 8.89277 11.448 8.80157C11.3528 8.71038 11.24 8.63902 11.1163 8.59175C10.9927 8.54447 10.8607 8.52224 10.7282 8.52636H3.72838C3.59817 8.52546 3.46908 8.55014 3.3486 8.59898C3.22813 8.64782 3.11867 8.71984 3.0266 8.81085C2.93452 8.90187 2.86167 9.01006 2.81226 9.12915C2.76285 9.24824 2.73788 9.37584 2.73879 9.50455C2.74359 9.76241 2.84942 10.0084 3.034 10.1906C3.21858 10.3729 3.46751 10.4773 3.72838 10.4818Z"
-                      fill="#333333"
-                    />
-                    <path
-                      d="M9.89078 13.2649C9.89169 13.1362 9.86672 13.0086 9.81731 12.8895C9.7679 12.7704 9.69505 12.6622 9.60297 12.5711C9.5109 12.4801 9.40144 12.4081 9.28097 12.3592C9.16049 12.3104 9.0314 12.2857 8.90119 12.2866H1.86403C1.73151 12.2825 1.59951 12.3047 1.47586 12.352C1.35221 12.3993 1.23943 12.4706 1.14422 12.5619C1.049 12.6531 0.9733 12.7622 0.921601 12.8829C0.869902 13.0036 0.843262 13.1333 0.843262 13.2644C0.843262 13.3955 0.869902 13.5252 0.921601 13.6459C0.9733 13.7666 1.049 13.8758 1.14422 13.967C1.23943 14.0582 1.35221 14.1295 1.47586 14.1768C1.59951 14.2241 1.73151 14.2463 1.86403 14.2422H8.86385C8.99756 14.2502 9.13148 14.2306 9.25716 14.1848C9.38285 14.139 9.49756 14.068 9.59407 13.9761C9.69057 13.8843 9.76676 13.7737 9.81783 13.6513C9.8689 13.5288 9.89374 13.3973 9.89078 13.2649Z"
-                      fill="#333333"
-                    />
-                  </svg>
-                </span>
-                <div className="media-body">
-                  <h4 className="media-title">Free shipping</h4>
-                  <p className="media-content">
-                    All purchases over $199 are eligible forfree shipping via USPS First ClassName Mail.
-                  </p>
-                </div>
-              </div>
-            </li>
-            <li className="benefit-item col-3 col-sm-12">
-              <div className="media media-2 media-vertical">
-                <span className="media-img benefit-icon">
-                  <svg
-                    width="35"
-                    height="37"
-                    viewBox="0 0 35 37"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      d="M2.90551 33.771C2.7364 33.7699 2.57288 33.7069 2.44276 33.5928C2.31263 33.4786 2.22395 33.3203 2.19181 33.1448C2.15966 32.9693 2.18603 32.7874 2.26644 32.6302C2.34684 32.4729 2.47631 32.35 2.63281 32.2822C2.78931 32.2145 2.96317 32.2062 3.12482 32.2587C3.28646 32.3113 3.42589 32.4214 3.51939 32.5703C3.61289 32.7193 3.65467 32.8979 3.63763 33.0757C3.62059 33.2536 3.54579 33.4197 3.42594 33.5458C3.35754 33.6178 3.27633 33.6747 3.18701 33.7134C3.09769 33.752 3.00202 33.7716 2.90551 33.771Z"
-                      fill="#333333"
-                    />
-                    <path
-                      d="M6.76666 37.0002H0V35.1481H5.01565V25.6567H0V23.8057H6.76666V37.0002Z"
-                      fill="#333333"
-                    />
-                    <path
-                      d="M17.1569 35.1105H7.92505V25.7156L8.0265 25.5107C8.5954 24.456 9.44968 23.6052 10.4795 23.0676C11.4459 22.5767 12.4997 22.3074 13.5715 22.2777C14.6433 22.2479 15.709 22.4583 16.6984 22.895C17.3355 23.1703 17.9574 23.4836 18.561 23.8334C19.0571 24.1079 19.5258 24.3696 20.0107 24.5841C20.2149 24.6632 20.4241 24.7274 20.6366 24.7761C21.0776 24.8214 21.5022 24.9764 21.8757 25.2284C22.2491 25.4804 22.5608 25.8223 22.7853 26.2261C23.0218 26.8651 23.1122 27.5541 23.0491 28.237L27.4601 28.1705L29.7478 25.4893C30.1722 24.9904 30.7668 24.6902 31.4006 24.6547C32.0344 24.6192 32.6557 24.8513 33.1276 25.3C33.5995 25.7487 33.8835 26.3772 33.9171 27.0473C33.9507 27.7173 33.7311 28.374 33.3066 28.8729L29.8574 32.9387L17.1569 35.1105ZM9.68215 33.2583H17.0169L28.9677 31.2206L32.002 27.6342C32.1144 27.5003 32.172 27.3249 32.1625 27.1462C32.1529 26.9675 32.0769 26.8001 31.951 26.6804C31.8251 26.5607 31.6595 26.4986 31.4904 26.5075C31.3213 26.5164 31.1624 26.5957 31.0484 26.728L28.2494 30.0044L17.4065 30.1781L17.3811 28.3271L21.3042 28.2659C21.3576 27.8309 21.3192 27.3888 21.1916 26.9714C21.1246 26.8106 20.5413 26.6583 20.2278 26.5768C19.9253 26.5067 19.6285 26.4116 19.3401 26.2926C18.7882 26.0459 18.2648 25.7563 17.7585 25.4753C17.213 25.1585 16.6511 24.8742 16.0755 24.6238C15.3174 24.2821 14.4996 24.1137 13.6756 24.1294C12.8517 24.1451 12.0401 24.3446 11.2942 24.715C10.6459 25.055 10.0917 25.5663 9.68519 26.1993L9.68215 33.2583Z"
-                      fill="#333333"
-                    />
-                    <path
-                      d="M35 4.13328V2.16424L29.4203 0L18.6444 5.85673H9.34448V19.6401H30.6945V15.7363C30.9847 15.747 31.2718 15.7588 31.5518 15.7803C31.7273 15.791 31.9018 15.7985 32.0773 15.7985C33.074 15.7905 34.0622 15.6041 35 15.2472V13.2685L34.9757 13.2793C33.9377 13.7709 32.8044 13.9955 31.6684 13.9345C31.3499 13.912 31.0283 13.8981 30.6955 13.8863C30.1355 13.8713 29.5583 13.8756 28.9455 13.897C28.1025 13.927 27.1986 13.9967 26.2125 14.104L25.1341 14.222C24.865 14.2612 24.5922 14.1859 24.3757 14.0125C24.1592 13.8391 24.0167 13.5819 23.9796 13.2975C23.9425 13.0131 24.0137 12.7247 24.1777 12.4958C24.3417 12.2669 24.585 12.1163 24.8541 12.077L28.9354 11.3263C28.937 11.327 28.9387 11.3274 28.9405 11.3274C28.9422 11.3274 28.9439 11.327 28.9455 11.3263C29.3037 11.2611 29.6447 11.1164 29.9462 10.9018C30.2476 10.6872 30.5027 10.4075 30.6945 10.0812C30.9424 9.66389 31.0802 9.18406 31.0934 8.69151C31.1067 8.19895 30.995 7.71161 30.7699 7.2801C30.5449 6.84858 30.2147 6.48862 29.8136 6.23757C29.4125 5.98652 28.9551 5.85354 28.489 5.85244H22.469L29.54 2.01624L35 4.13328ZM28.49 7.70674C28.6512 7.70499 28.8094 7.75267 28.9455 7.84401C29.069 7.9261 29.1707 8.03994 29.2412 8.17489C29.3117 8.30983 29.3487 8.46148 29.3487 8.61565C29.3487 8.76982 29.3117 8.92147 29.2412 9.05642C29.1707 9.19137 29.069 9.30521 28.9455 9.38729C28.8526 9.44866 28.7492 9.4902 28.6412 9.50955L24.5599 10.2603C23.8745 10.384 23.2581 10.7755 22.8328 11.357C22.4076 11.9386 22.2046 12.6679 22.2641 13.4003C22.3236 14.1326 22.6412 14.8146 23.154 15.3111C23.6668 15.8075 24.3373 16.0822 25.0326 16.0805C25.1262 16.081 25.2196 16.076 25.3126 16.0655C25.3126 16.0655 28.1613 15.7845 28.9496 15.7545V17.7922H11.0945V7.71103L28.49 7.70674Z"
-                      fill="#333333"
-                    />
-                    <path d="M21.0092 11.8228H19.2593V13.6729H21.0092V11.8228Z" fill="#333333" />
-                  </svg>
-                </span>
-                <div className="media-body">
-                  <h4 className="media-title">Easy Payments</h4>
-                  <p className="media-content">
-                    All payments are processed instantlyover a secure payment protocol.
-                  </p>
-                </div>
-              </div>
-            </li>
-            <li className="benefit-item col-3 col-sm-12">
-              <div className="media media-2 media-vertical">
-                <span className="media-img benefit-icon">
-                  <svg
-                    width="29"
-                    height="31"
-                    viewBox="0 0 29 31"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      d="M27.546 3.99844L14.5018 0.0450776C14.3078 -0.0150259 14.1013 -0.0150259 13.9073 0.0450776L0.86503 3.99844C0.650172 4.0662 0.462005 4.20437 0.32847 4.39246C0.194936 4.58054 0.123154 4.8085 0.123783 5.04248V16.9137C0.123783 27.2427 13.3914 30.8204 13.9597 30.9672C14.1311 31.0126 14.3109 31.0108 14.4814 30.9621C15.2955 30.7495 28.297 27.1495 28.297 16.9137V5.04248C28.297 4.8072 28.2238 4.57827 28.0884 4.39002C27.953 4.20178 27.7627 4.06438 27.546 3.99844ZM26.2102 16.9137C26.2102 24.8539 15.9639 28.2796 14.2104 28.786C12.4617 28.2796 2.21054 24.8701 2.21054 16.9137V5.86374L14.2104 2.22024L26.2102 5.85867V16.9137Z"
-                      fill="#333333"
-                    />
-                    <path
-                      d="M13.1659 8.70233V9.84257C12.3907 9.97748 11.6926 10.4114 11.2102 11.0582C10.7278 11.705 10.4965 12.5172 10.5622 13.3336C10.6279 14.1499 10.9859 14.9106 11.5649 15.4646C12.1439 16.0185 12.9017 16.3251 13.6876 16.3235H15.2526C15.5294 16.3235 15.7947 16.4381 15.9904 16.642C16.1861 16.846 16.296 17.1226 16.296 17.4111C16.296 17.6995 16.1861 17.9761 15.9904 18.1801C15.7947 18.3841 15.5294 18.4987 15.2526 18.4987H12.1235C11.8468 18.4987 11.5814 18.6132 11.3857 18.8172C11.1901 19.0212 11.0801 19.2978 11.0801 19.5862C11.0801 19.8747 11.1901 20.1513 11.3857 20.3553C11.5814 20.5592 11.8468 20.6738 12.1235 20.6738H13.1659V21.7614C13.1659 22.0498 13.2758 22.3265 13.4715 22.5304C13.6672 22.7344 13.9326 22.849 14.2093 22.849C14.486 22.849 14.7514 22.7344 14.947 22.5304C15.1427 22.3265 15.2526 22.0498 15.2526 21.7614V20.6667C16.0828 20.6667 16.879 20.323 17.466 19.7111C18.053 19.0992 18.3828 18.2693 18.3828 17.404C18.3828 16.5387 18.053 15.7088 17.466 15.0969C16.879 14.485 16.0828 14.1412 15.2526 14.1412H13.6886C13.4118 14.1412 13.1465 14.0267 12.9508 13.8227C12.7551 13.6187 12.6452 13.3421 12.6452 13.0537C12.6452 12.7652 12.7551 12.4886 12.9508 12.2846C13.1465 12.0807 13.4118 11.9661 13.6886 11.9661H16.8187C17.0954 11.9661 17.3608 11.8515 17.5564 11.6475C17.7521 11.4436 17.862 11.1669 17.862 10.8785C17.862 10.5901 17.7521 10.3134 17.5564 10.1095C17.3608 9.9055 17.0954 9.79092 16.8187 9.79092H15.2536V8.70233C15.2536 8.5595 15.2266 8.41808 15.1742 8.28613C15.1218 8.15418 15.0449 8.03428 14.948 7.93329C14.8511 7.8323 14.7361 7.75219 14.6095 7.69753C14.4829 7.64288 14.3473 7.61475 14.2102 7.61475C14.0732 7.61475 13.9375 7.64288 13.811 7.69753C13.6844 7.75219 13.5694 7.8323 13.4725 7.93329C13.3756 8.03428 13.2987 8.15418 13.2463 8.28613C13.1939 8.41808 13.1669 8.5595 13.1669 8.70233H13.1659Z"
-                      fill="#333333"
-                    />
-                  </svg>
-                </span>
-                <div className="media-body">
-                  <h4 className="media-title">Money-Back Guarantee</h4>
-                  <p className="media-content">
-                    If an item arrived damaged or you'vechanged your mind, you can send itback for a
-                    full refund.
-                  </p>
-                </div>
-              </div>
-            </li>
-            <li className="benefit-item col-3 col-sm-12">
-              <div className="media media-2 media-vertical">
-                <span className="media-img benefit-icon">
-                  <svg
-                    width="33"
-                    height="32"
-                    viewBox="0 0 33 32"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      d="M9.56732 21.6376C10.1057 21.6376 10.5422 21.2111 10.5422 20.685C10.5422 20.1589 10.1057 19.7324 9.56732 19.7324C9.02889 19.7324 8.59241 20.1589 8.59241 20.685C8.59241 21.2111 9.02889 21.6376 9.56732 21.6376Z"
-                      fill="#333333"
-                    />
-                    <path
-                      d="M32.5805 0H10.7208V10.6421H0.208984V32H22.0677V21.3579H32.5805V0ZM30.88 4.55072L22.0696 13.1593V10.6421H18.8292L27.9429 1.73721L27.8673 1.66337H30.88V4.55165V4.55072ZM20.3663 21.3044C18.9058 21.0644 16.6385 20.794 15.8232 23.4782C15.1906 25.06 14.3267 26.5437 13.2583 27.8831C12.6568 28.6472 12.1368 29.4693 11.7061 30.3366H9.01086C9.00483 29.063 9.43442 27.8239 10.2314 26.816C10.6041 26.4048 11.0363 26.049 11.5143 25.7601C12.3108 25.3147 12.9887 24.6922 13.4931 23.9431C13.9975 23.194 14.3141 22.3394 14.4174 21.4484C14.7056 19.857 14.5084 18.218 13.8506 16.7352C13.2406 15.3395 13.03 13.8081 13.2413 12.3045H15.7334C15.8033 12.6442 15.8638 12.9959 15.9224 13.3669C16.2795 15.4679 16.7235 18.082 19.5812 19.5183C19.7409 19.5986 19.8448 19.654 19.9109 19.6918C20.0381 19.8209 20.1934 19.9203 20.3653 19.9826V21.3044H20.3663ZM6.53761 22.7813C6.4483 21.9338 6.50514 21.0776 6.70577 20.2484C6.94722 19.4208 7.34168 18.6434 7.86965 17.9546C9.09965 16.3154 9.76162 14.3352 9.75907 12.3027H11.5314C11.3276 14.0037 11.5759 15.7274 12.2522 17.3075C12.8084 18.5252 12.9782 19.8779 12.7396 21.1909C12.6709 21.859 12.4349 22.5003 12.0525 23.0583C11.6701 23.6162 11.1531 24.0737 10.547 24.3902C9.93031 24.7676 9.37602 25.2345 8.90316 25.7748C7.86375 27.0769 7.30245 28.6826 7.30944 30.3339H5.53149C6.35272 27.9018 6.69413 25.3396 6.53761 22.7832V22.7813ZM20.3663 18.0414H20.3597C18.2624 16.9872 17.9412 15.0968 17.6011 13.0956C17.5567 12.8343 17.5067 12.5694 17.4604 12.3054H20.3663V18.0414ZM12.4232 9.16791L20.1065 1.6606H25.6076L16.4183 10.6393H12.4232V9.16791ZM17.7003 1.66245L12.4232 6.8187V1.66245H17.7003ZM4.55466 12.3045C4.155 13.7764 3.6236 15.2112 2.9666 16.5921C2.61989 17.3971 2.25806 18.238 1.91135 19.1094V12.3045H4.55466ZM1.90947 24.9939C2.47567 22.3193 3.35599 19.7175 4.53293 17.2401C5.2995 15.6638 5.88531 14.0094 6.2797 12.3073H8.05576C8.05746 14.0123 7.49402 15.672 6.44975 17.038C5.81462 17.8801 5.34179 18.8281 5.05441 19.8358C4.80867 20.8275 4.73396 21.8523 4.83335 22.8681C4.97805 25.4035 4.61149 27.9421 3.75449 30.3394H1.91135V24.9939H1.90947ZM13.6173 30.3366C13.9092 29.8751 14.3154 29.3046 14.6508 28.8366C15.8305 27.3551 16.7751 25.7083 17.4528 23.9518C17.8354 22.6918 18.5241 22.6826 20.252 22.9724L20.3653 22.9909V30.3366H13.6173ZM22.0677 15.5103L30.8781 6.90178V12.5057L23.5197 19.6955H22.0677V15.5103ZM25.9259 19.6955L30.8781 14.8568V19.6955H25.9259Z"
-                      fill="#333333"
-                    />
-                  </svg>
-                </span>
-                <div className="media-body">
-                  <h4 className="media-title">Finest Quality</h4>
-                  <p className="media-content">
-                    Designed to last, each of our products hasbeen crafted with the finest
-                    materials.
-                  </p>
-                </div>
-              </div>
-            </li>
-          </ul>
-        </div>
-      </section>
-      <section className="section-products page-section">
-        <div className="container">
-          <div className="section-heading">
-            <h3 className="section-title text-center">Products in today</h3>
-          </div>
-          <ProductList data={productData}/>
-        </div>
-      </section>
-      <section className="section-subcribe">
-        <div className="container">
-          <div className="subcribe row">
-            <div className="col-6 col-sm-12">
-              <h2 className="subcribe-title">
-                Subscribe to our newsletter andreceive exclusive offers every week
-              </h2>
-            </div>
-            <div className="col-6 col-sm-12">
-              <form className="subcribe-form" action="">
-                <input className="form-control" type="email" placeholder="Enter your email" />
-                <button className="subcribe-btn btn btn-primary">SUBCRIBE</button>
-              </form>
-            </div>
-          </div>
-        </div>
-      </section>
-    </main>
-  )
-}
-
-export default Home;
