+++ e-shop/src/app/pages/home/components/Banner/Offer.tsx
@@ -1,16 +1,15 @@
-const Policy = ({title, content, icon}: any) => {
+const Offer = ({title, desc, icon}: any) => {
   return (
-    <li className="offer-item">
-      <div className="media media-sm">
+    <div className="media media-sm">
         <span className="media-img offer-icon">
           <img src={icon}/>
         </span>
         <div className="media-body">
           <h4 className="media-title">{title}</h4>
-          <p className="media-content">{content}</p>
+          <p className="media-content">{desc}</p>
         </div>
       </div>
-    </li>
   )
 };
-export default Policy;
+
+export default Offer;
