--- e-shop/src/app/shared/components/Button.tsx
@@ -1,7 +1,0 @@
-const Button = ({children} : any) => {
-  return (
-    <button>{children}</button>
-  )
-}
-
-export default Button;
