+++ e-shop/src/app/shared/components/Product.tsx
