+++ e-shop/src/app/shared/components/layout/Footer.tsx
@@ -1,4 +1,4 @@
-function Footer() {
+const Footer = () => {
   return (
     <footer>
       <div className="container">
