+++ e-shop/src/app/shared/components/layout/Header.tsx
@@ -1,4 +1,4 @@
-function Header() {
+const Header = () => {
   return (
     <header>
       <div className="container">
