+++ e-shop/src/app/shared/constants/offer-data.ts
@@ -1,0 +1,24 @@
+import Icon from '../../assets/icons';
+
+const offerData = [
+  {
+    id: 1,
+    icon: Icon.Shipping,
+    title: 'Free Shipping',
+    desc: 'On purchases over $199'
+  },
+  {
+    id: 2,
+    icon: Icon.Happy,
+    title: '99% Satisfied Customers',
+    desc: 'Our clients opinions speak for themselves'
+  },
+  {
+    id: 3,
+    icon: Icon.Guarantee,
+    title: 'Originality Guaranteed',
+    desc: '30 days warranty for each product from our store'
+  },
+];
+
+export default offerData;
