+++ e-shop/src/app/shared/constants/product-data.ts
